+++ src/App/App.php
@@ -11,7 +11,7 @@
  * @author Jörn Heinemann <joernheinemann@gmx.de>
  */
 
-/** @version 1.6.0 */
+/** @version 1.6.1 */
 
 namespace Gram\App;