+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 	- bei true -> render exception
 	- bei false -> nur Response mit 500 zurück geben
 - Bei 404 bzw. 405 wird nun der richtige Statuscode bei Exceptions ausgegeben, sollte kein Handler gesetzt sein
+- Resolver müssen nun bei resolve() ein array mit den Paramtern bekommen
+
 
 # 1.5.x